Implement patient data fetching and vital signs extraction in Dashboard and ProfileMedHistory components
