+++ components/profile-med-history.tsx
@@ -3,6 +3,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { EmergencyContact } from './emergency-contact';
 import { Patient } from '@/lib/types/patient';
 import fetchPatientData from '@/lib/fetchPatientData';
+import { extractVitalSigns } from '@/lib/extractVitalSigns';
 
 const excludedHistoryKeys = ['_id', 'patient_name', 'date_of_birth', 'gender'];
 
@@ -89,6 +90,7 @@ const ProfileMedHistory: React.FC = () => {
       try {
         const mappedPatient = await fetchPatientData();
         setPatient(mappedPatient);
+        console.log(extractVitalSigns(mappedPatient.history["Vital Signs"]))
       } catch (err) {
         setError('Failed to fetch patient data');
       } finally {
@@ -156,68 +158,6 @@ const ProfileMedHistory: React.FC = () => {
     </Card>
   );
 };
-
-// Define a type for the vital signs entries.
-export type VitalSigns = {
-    date: Date; // raw date; you can format it when displaying
-    bodyHeight: number;
-    painSeverity: number;
-    bodyWeight: number;
-    bmi: number;
-    heartRate: number;
-    respiratoryRate: number;
-  };
-  
-  // A helper that extracts a number from a string (ignoring units and tags)
-  const extractNumber = (value: string): number => {
-    // Use parseFloat to extract the first number it finds.
-    return parseFloat(value);
-  };
-  
-  export function extractVitalSigns(vitalSignsData: Record<string, any>): VitalSigns[] {
-    // vitalSignsData is expected to be an object where each key is a date string.
-    const results: VitalSigns[] = Object.entries(vitalSignsData).map(([dateKey, records]) => {
-      let bodyHeight = NaN,
-        painSeverity = NaN,
-        bodyWeight = NaN,
-        bmi = NaN,
-        heartRate = NaN,
-        respiratoryRate = NaN;
-        
-      // records is expected to be an array of objects
-      for (const record of records) {
-        const desc = record.Description.toLowerCase();
-        const val = record.Value;
-        const num = extractNumber(val);
-        if (desc.includes("body height")) {
-          bodyHeight = num;
-        } else if (desc.includes("pain severity")) {
-          painSeverity = num;
-        } else if (desc.includes("body weight")) {
-          bodyWeight = num;
-        } else if (desc.includes("body mass index")) {
-          bmi = num;
-        } else if (desc.includes("heart rate")) {
-          heartRate = num;
-        } else if (desc.includes("respiratory rate")) {
-          respiratoryRate = num;
-        }
-      }
-      return {
-        date: new Date(dateKey),
-        bodyHeight,
-        painSeverity,
-        bodyWeight,
-        bmi,
-        heartRate,
-        respiratoryRate,
-      };
-    });
-  
-    // Sort the array in descending order (most recent date first)
-    results.sort((a, b) => b.date.getTime() - a.date.getTime());
-    return results;
-  }
   
 
 export default ProfileMedHistory;
